refactor depsearch
